I've tried changing the karma configuration to use the karma-webpack plugin. When the webpack configuration is included, and the bundle is provided as one of the  files in the files: array, it will load. However, there are 172 Karma tests that should run. Several will be skipped because they are commented with XIT so they are skipped.
However, what happens is that only 38 tests are run, most are skipped, and only 6 are executed. When this test configuration is run, there are no errors that are thrown, but I am not conviced that any of the test cases are actually being run. There is a section in the Karma configuration for preprocessor. The test cases in /src/test/frontend/spec/*.js should all be exectued.

The karma-webpack preprocessor is supposed to build each of those files into their own small bundles and test them serially in memory. I am not sure if those test cases are being picked up or if the only the fixtures are.


The second part of what I did, and the current state of the test, is to revert those changes I had made to include the karma-webpack plugin. This returned me to the original state where the variables aren't being resolved correctly.
Karma/Jasmine expects several global variables that aren't being called correctly within the TDAR modules. Specifically, I think this is having to do with issues with the way Karma is loading fixures as JQUery objects, and the jQuery object within the bundle is different.
There's several failures where a fixure is definitely correct and is parsed correctly, but the module (TDAR.*) can't find the objects to manipulate. In debugging, Both the Jasmine script and the TDAR component both expose Jquery, so it is not an issue with it being null.

I tried commenting out the files in the files array within the Karma configuration because many of those files are already included in the bundle. This didn't seem to resolve anything though. I had also tried exposing additional variables as global within the webpack.config.js ProvidePlugin section, but I did not see any change.