The jasmine-jquery jasmine.getFixtures().fixturesPath doesn't default to the correct location. I am not sure if this was removed form the config earlier, or if something changed to keep it from working. The fix at the moment is to wrap beforeEach(){} to call and set the fixturesPath.
The main issue I am having is not with the variables not being found ,but the DOM differences with JQuery selectors. I believe Jasmine/Karma has one version of jQuery loaded, and the Webpack bundle uses another. The fixures are loaded into one JQuery dom, and the TDAR modules to check a different dom with its embedded JQuery.

I want to remove JQuery from the Webpack bundle so it doesn't embed that version into it.