temporarily removing rollbar.js (TDAR-6234) (see notes)
Per TDAR-6234,  it looks like IE11 is crashing when the browser
attempts to render dataset page when user is logged in. IE11 renders
these same pages without crashing when I omit the Rollbar.js include
snippet.   I suspect that there is an underlying issue at play, and
that rollbar + datatables.net widget is not the root cause. In the
short term, I'm disabling rollbar so that datatables render
properly on IE11 - we have governement clients that are only allowed
to use IE, therefore they effectively cannot use tDAR as long as
this issue exists.