adding singleToken parameter to DataValue (see notes)
I added a 'singletoken' parameter to the DataValue as a way to
signal when to query on the 'value' field instead of both
'value' and 'value_phrase'.

HOWEVER, shortly before this commit, I discovered that the
query snippet generated by DataValueQueryPart appears to
be malformed.  Specifically, the querypart generates the
'inner' query in the 'v' local parameter. I don't think
this is valid, as the parsed query does not utilize
the tokenqueryparser as it should.   I think that perhaps
the 'v' local parameter forces the query to apply to the
field specified by the local parameter section, and because
we do not specify a field, the query applies to the default
field ('_text_' maybe?).
anyway, using the following form appears to work:

{!join fromIndex=dataMappings from=id to=id}
value:(St.\ Louis\ Art\ Museum) AND columnId:(32655)