paring down vue-advanced-search into vue-dataset-search (see notes)
The dataset search page no longer requires the full range of
serach controls that you'd see on the advanced search page,
and the complexity was making it difficult for me to
implement serialization/deserialization, so I'm paring
it down to the minimal set of functionality and and renaming
it to vue-dataset-search to avoid confusiton.  Once this
component actually works, it would be a good idea to revisit
an advanced search vue control.