show bookmark tool for any authenticated user  (fixes TDAR-6368, see notes)
Looks like this feature was gated back in Jan 2018 such that
users must have edit rights on the resource in question,
however,  it is not not otherwise necessary restrict which
resources a user can bookmark.   I can't see any commit notes
or tickets that reference the change so I am re-adding it.