initial support for ESRI worldmap (see notes)
I've managed to swap out the OSM baselayer for the ESRI version
by adding an URL template similar to what we use for OSM and
Mapbox.  note that this URL template is not mentioned in the
ESRI github sites or their documentation site. Instead I found
this example URL template somewhere online - I *think* that this
url format in the "leaflet extras" documentation plugin.

According to the openstreetmap wiki,  the ESRI worldmap is free
to use (and does not even require attribution).  That said, given
the sparse, old documentation and also the fact that we are using
a very old version of leaflet, I ultimately think that the we
should refactor all of our map code to run under the latest
version of leaflet (currently at v1.7)