adding workaround for selenium httpclient issue (see notes)
An updated version of the chromium webdriver fails due to some
issue with the httpclient that selenium uses.  The gist
is that chromedriveer refuses connections from selenium because
something about the request headers makes the chromedriver
rejects the connection as "remote" (or somesuch).  Work-around
was to allow chromedriver to accept remote connections.  This
doesn't affect the security of digital antiquity's filrewalled
build server, but worth noting.

 More info:
- https://stackoverflow.com/questions/75678572/java-io-ioexception-invalid-status-code-403-text-forbidden
- https://github.com/SeleniumHQ/selenium/issues/11750