possible fix for collection autocomplete issue (see notes)
looks like the collection autocomplete is missing an "autocompletename"
value, which is used by applyDataElements.  We use this attribute on the
collection autocomplete on the "collection export" page, which does not
show the same buggy behavior.

Note this is not a complete fix.  On the server end, the system should
add a resource to a collection even if the collection ID is missing, if
the collection matches the name of an existing collection *and* if the
user has "add to collection" permissions or better.  I think the issue
is that "add to collection" permission is a "new" permission that was
created after the the "add to collection" permission, and we didn't
update the server logic accordingly.